+++ src/hypernova/mod.rs
@@ -6,7 +6,9 @@ use crate::frontend::{num::AllocatedNum, ConstraintSystem, SynthesisError};
 mod augmented_circuit;
 pub mod error;
 pub mod nifs;
+mod poseidon_sumcheck;
 pub mod rs;
+
 /// Step circuit used for Hypernova
 pub trait StepCircuit<F: PrimeField> {
   /// Arity of the circuit. This is needed to build the public parameters
